requirements.txt
torch==1.13.1
numpy==1.22.3
scikit_learn==1.1.1
tensorboard==2.9.0